Consolidate toolbar and treeview components
