+++ src/components/BrokerQueueTreeView/styles.module.css
@@ -1,0 +1,17 @@
+.container {
+  width: 100%;
+}
+
+.toolbar {
+  padding: 0px;
+  border: unset;
+  border-radius: unset;
+  user-select: none;
+}
+
+.tree {
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  user-select: none;
+}
