--- src/components/Toolbar/styles.module.css
@@ -1,4 +1,0 @@
-.toolbar {
-  padding: 2px;
-  border-radius: unset;
-}
