--- src/components/TreeView/styles.module.css
@@ -1,10 +1,0 @@
-.scrollPanel {
-  width: 100%;
-}
-
-.tree {
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  user-select: none;
-}
